+++ certbot-dns-rfc2136/certbot_dns_rfc2136/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_rfc2136.dns_rfc2136` plugin.
