+++ certbot-dns-rfc2136/docs/api/dns_rfc2136.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_rfc2136.dns_rfc2136`
+:mod:`certbot_dns_rfc2136._internal.dns_rfc2136`
 --------------------------------------
 
-.. automodule:: certbot_dns_rfc2136.dns_rfc2136
+.. automodule:: certbot_dns_rfc2136._internal.dns_rfc2136
    :members:
